Divided Token class
