+++ backend/users/backend/common/services/jwt/token.py
@@ -6,27 +6,12 @@ from base.tokens.token import BaseToken
 from common.services.jwt.exceptions import PayloadError, TokenExpired
 from common.services.jwt.mixins import JWTMixin
 from common.services.jwt.users_payload import PayloadFactory
-from config.settings import redis_config
+from common.services.jwt.redis import RedisAccessToken, RedisBanRefreshToken
 from django.conf import settings
 from django.utils import timezone
-from utils.db.redis_client import RedisAccessClient, RedisClient, RedisRefreshClient
 
 
 class Token(BaseToken, JWTMixin):
-    redis_access_client = RedisAccessClient(
-        host=redis_config.REDIS_SHARED_HOST,
-        port=redis_config.REDIS_SHARED_PORT,
-        db=redis_config.REDIS_ACCESS_DB,
-        password=redis_config.REDIS_SHARED_PASSWORD,
-    )
-
-    redis_refresh_client = RedisRefreshClient(
-        host=redis_config.REDIS_SHARED_HOST,
-        port=redis_config.REDIS_SHARED_PORT,
-        db=redis_config.REDIS_REFRESH_DB,
-        password=redis_config.REDIS_SHARED_PASSWORD,
-    )
-
     @staticmethod
     def validate_user(user: BaseAbstractUser):
         if user.is_active is not True:
@@ -80,7 +65,7 @@ class Token(BaseToken, JWTMixin):
             **default_payload,
         }
         access_token = self._encode(payload)
-        self.__add_access_to_redis(token=access_token, value=redis_payload)
+        RedisAccessToken.add_access_to_redis(token=access_token, value=redis_payload)
         return access_token
 
     def generate_refresh_token(self, data: dict) -> str:
@@ -115,7 +100,7 @@ class Token(BaseToken, JWTMixin):
             **default_payload,
         }
         access_token = self._encode(payload)
-        self.__add_access_to_redis(token=access_token, value=redis_payload)
+        RedisAccessToken.add_access_to_redis(token=access_token, value=redis_payload)
         return access_token
 
     def generate_refresh_token_for_user(self, user: BaseAbstractUser) -> str:
@@ -160,21 +145,9 @@ class Token(BaseToken, JWTMixin):
 
     def check_signature(self, token: str) -> None:
         self._decode(token)
-
-    @staticmethod
-    def __add_token_to_redis(redis_client: RedisClient, token: str, value: dict):
-        redis_client.add_token(token=token, value=value)
-
-    @classmethod
-    def __add_access_to_redis(cls, token: str, value: dict):
-        cls.__add_token_to_redis(redis_client=cls.redis_access_client, token=token, value=value)
-
-    @classmethod
-    def add_refresh_to_redis(cls, token: str, value: dict = None):
-        cls.__add_token_to_redis(redis_client=cls.redis_refresh_client, token=token, value=value)
-
-    def get_access_data(self, token: str) -> dict | None:
-        return self.redis_access_client.is_token_exist(token)
-
-    def get_refresh_data(self, token: str) -> dict | None:
-        return self.redis_refresh_client.is_token_exist(token)
+    
+    # for backward compatibility
+    redis_refresh_client = RedisBanRefreshToken.redis_refresh_client
+    add_refresh_to_redis = RedisBanRefreshToken().add_refresh_to_redis
+    get_access_data = RedisAccessToken().get_access_data
+    get_refresh_data = RedisBanRefreshToken().get_refresh_data
